+++ lua/terrortown/lang/en/deadringer.lua
@@ -1,0 +1,40 @@
+local L = LANG.GetLanguageTableReference('en')
+
+L.deadringer_name = 'Dead Ringer'
+L.deadringer_desc = 'A device that fakes your death when you take damage, giving you a chance to escape.'
+
+L.deadringer_primary = 'Activate'
+L.deadringer_secondary = 'Deactivate'
+
+L.label_deadringer_cloak_cooldown = 'Cloak Cooldown'
+L.help_deadringer_cloak_cooldown = 'The time you have to wait before you can use the Dead Ringer again.'
+
+L.label_deadringer_cloak_duration = 'Cloak Duration'
+L.help_deadringer_cloak_duration = 'The time you are invisible after using the Dead Ringer.'
+
+L.label_deadringer_cloak_reuse = 'Cloak Reuse'
+L.help_deadringer_cloak_reuse = 'Whether or not the Dead Ringer will convert unused cloak time into cooldown reduction.'
+
+L.label_deadringer_cloak_transparency = 'Cloak Transparency'
+L.help_deadringer_cloak_transparency = 'The transparency of the player while cloaked.'
+
+L.label_deadringer_cloak_targetid = 'Cloak Target ID'
+L.help_deadringer_cloak_targetid = 'Whether or not the Dead Ringer will show/hide the target ID while cloaked.'
+
+L.label_deadringer_damage_reduction = 'Damage Reduction'
+L.help_deadringer_damage_reduction = 'The percentage of damage you take while cloaked.'
+
+L.label_deadringer_damage_threshold = 'Damage Threshold'
+L.help_deadringer_damage_threshold = 'The amount of damage you need to take in a single hit to trigger the Dead Ringer.'
+
+L.label_deadringer_damage_reduction_initial = 'Initial Damage Reduction'
+L.help_deadringer_damage_reduction_initial = 'The percentage of damage you receive when the Dead Ringer is triggered.'
+
+L.label_deadringer_damage_reduction_time = 'Damage Reduction Time'
+L.help_deadringer_damage_reduction_time = 'The time you receive reduced damage after the Dead Ringer is triggered.'
+
+L.label_deadringer_corpse_mode = 'Corpse Mode'
+L.help_deadringer_corpse_mode = 'Whether or not the Dead Ringer will show the real role of the player on the corpse. (0 = fake, 1 = real, 2 = innocent)'
+
+L.label_deadringer_corpse_confirm = 'Corpse Confirm'
+L.help_deadringer_corpse_confirm = 'Whether or not the Dead Ringer will confirm the death of the player on the corpse.'